+++ data/jimmy
@@ -1,0 +1,5 @@
+T | 0 | todo 1
+D | 0 | todo | 19-01-2025
+E | 0 | todo | 19-01-2025 | 20-01-2025
+T | 0 | todo1
+D | 0 | deadline | 19-01-2025
