+++ src/main/java/Ui.java
@@ -1,0 +1,16 @@
+public class Ui {
+
+    /**
+     * Greets the user.
+     */
+    public void greetUser() {
+        System.out.println("Hello! I'm Jimmy\nWhat can I do for you?\n");
+    }
+
+    /**
+     * Exits the bot.
+     */
+    public void exit() {
+        System.out.println("Bye. Hope to see you again soon!");
+    }
+}
